Extract problem formatter to separate module
This is done in preparation to make Dantzig able to use different
solvers, not just HiGHS.
